Bump to 4.5.x-dev
This conflicts with an earlier commit (cfba4cb), but after talking to Robbie this new commit is consistent with how we handle other recipes (CWP and core). Both ways are viable, but we've implicitly standardised on this way (set minor release constraints in the main release branch, e.g. constraint to recipe:4.5.x-dev on installer:4.x-dev)