ENH Add and ignore public/_graphql
silverstripe/graphql v4 needs write access to this directory, so it may
be easier for some developers to have the directory there as soon as
they `composer create_project silverstripe/installer`.

Like `.graphql-generated`, best practice is to not commit this directory
to VCS, and instead let it be filled at deploy time.